Added skeleton functions for binary negation and binary multiplication.
Refactored code from a prior project for binary addition. Added validity checking on inputs so that it matched a particular regular expression.

Removed user inputting of results (BinaryAddition, BinaryMultiplication, BinaryNegation1, BinaryNegation2).

Added functionality to generate the results upon creating or editing an Expression.

Updated the layout to make it easier to navigate.