It turns out that the way the project was previously structured had major flaws.
Not all operations have the same inputs and output structures. Negation had one input of length 7 that result in one output of length 8. Addition has two inputs of length 7 that result in one output of length 8. Multiplication, on the other hand, would have two inputs of length 7 and one output of length 14.

Users might not want to do binary multiplication, addition and negation on the same set of inputs. It is not a good idea to do things that the user might not want to do. It is easier for each segment to do one particular thing than to have a table for all of the calculations in one place.

In order to deal with these issues, I decided to change the structure of the project. Instead of having one table that contains all of the user inputs and the appropriately generated outputs, I have created three models that correspond to the three different expressions that can be evaluated. More expressions (subtraction, division) might be added later, if I deem it necessary.

The next steps would be:

1) Write the binary multiplication function
2) Make the tables accessible from the header
3) Update the misc pages of the web application with relevant information.